Update cayenne-guide.html use Apache Cayenne in title
The ASF pattern is to use the proper name of the product (Apache Whatever) in the title of the documentation and at least the first use of the product name in the text. Once you have established the proper name, I think it is okay to just use Cayenne in the bulk of the documentation.